% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcICER.R
\name{calcICER}
\alias{calcICER}
\title{Calculate the Incremental Cost Effectivness Ratio}
\usage{
calcICER(e_int, e_base, c_int, c_base)
}
\arguments{
\item{e_int}{a single numeric value representing the effect (e.g. Total QALYs) in the intervention group.}

\item{e_base}{a single numeric value representing the effect (e.g. Total QALYs) in the base group.}

\item{c_int}{a single numeric value representing the cost (e.g. Total £) in the intervention group.}

\item{c_base}{a single numeric value representing the cost (e.g. Total £) in the base group.}
}
\value{
an single numeric value for the ICER.
}
\description{
Calculates the incremental effect and incremental costs of an intervention
compared to baseline and then uses the results to calculate the ICER.
}
\examples{

calcICER(e_int = 28.3, e_base = 22.5, c_int = 10000, c_base = 9200)

}
