% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_res_example}
\alias{df_res_example}
\title{Example Dataset for calcICER}
\format{
\subsection{\code{df_res_example}}{

A data frame with 1,000 rows and 4 columns:
\describe{
\item{e_base, e_int}{QALYs for base and intervention groups}
\item{c_base, c_int}{Costs for base and intervention groups}
...
}
}
}
\usage{
df_res_example
}
\description{
An example dataset on which to run calcICER
}
\keyword{datasets}
