% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mx.R
\name{get_mx}
\alias{get_mx}
\title{Calculate mortality rates (mx) for given ages and sexes}
\usage{
get_mx(v_ages, v_sex)
}
\arguments{
\item{v_ages}{A vector of ages.}

\item{v_sex}{A vector of sexes (either "M" or "F").}
}
\value{
A data frame with columns for age, sex, and mx.
}
\description{
This function takes two vectors, \code{v_ages} and \code{v_sex}, and calculates
mortality rates (mx) for the corresponding ages and sexes. The function returns
a data frame with columns for age, sex, and mx.
}
\examples{
get_mx(c(30, 40, 50), c("M", "F", "M"))

}
